chore: end_4

+++ install.sh
@@ -62,7 +62,7 @@ esac
 remove_bashcomments_emptylines ${DEPLISTFILE} ./cache/dependencies_stripped.conf
 readarray -t pkglist < ./cache/dependencies_stripped.conf
 
-# Use yay. Because paru do not support cleanbuild.
+# Use yay. Because paru does not support cleanbuild.
 # Also see https://wiki.hyprland.org/FAQ/#how-do-i-update
 if ! command -v yay >/dev/null 2>&1;then
   echo -e "\e[33m[$0]: \"yay\" not found.\e[0m"
@@ -81,24 +81,8 @@ if (( ${#pkglist[@]} != 0 )); then
 	fi
 fi
 
-# Convert old dependencies to non explicit dependencies so that they can be orphaned if not in meta packages 
-set-explicit-to-implicit() {
-	remove_bashcomments_emptylines ./scriptdata/previous_dependencies.conf ./cache/old_deps_stripped.conf
-	readarray -t old_deps_list < ./cache/old_deps_stripped.conf
-	pacman -Qeq > ./cache/pacman_explicit_packages
-	readarray -t explicitly_installed < ./cache/pacman_explicit_packages
-
-	echo "Attempting to set previously explicitly installed deps as implicit..."
-	for i in "${explicitly_installed[@]}"; do for j in "${old_deps_list[@]}"; do
-		[ "$i" = "$j" ] && yay -D --asdeps "$i"
-	done; done
-
-	return 0
-}
-
-$ask && echo "Attempt to set previously explicitly installed deps as implicit? "
-$ask && showfun set-explicit-to-implicit
-v set-explicit-to-implicit
+showfun handle-deprecated-dependencies
+v handle-deprecated-dependencies
 
 # https://github.com/end-4/dots-hyprland/issues/581
 # yay -Bi is kinda hit or miss, instead cd into the relevant directory and manually source and install deps
@@ -116,8 +100,9 @@ install-local-pkgbuild() {
 }
 
 # Install core dependencies from the meta-packages
-metapkgs=(./arch-packages/illogical-impulse-{audio,backlight,basic,fonts-themes,gnome,gtk,microtex,portal,python,screencapture,widgets})
-metapkgs+=(./arch-packages/illogical-impulse-ags)
+metapkgs=(./arch-packages/illogical-impulse-{audio,python,backlight,basic,fonts-themes,gnome,gtk,portal,screencapture,widgets})
+metapkgs+=(./arch-packages/illogical-impulse-agsv1)
+metapkgs+=(./arch-packages/illogical-impulse-hyprland)
 metapkgs+=(./arch-packages/illogical-impulse-oneui4-icons-git)
 [[ -f /usr/share/icons/Bibata-Modern-Classic/index.theme ]] || \
   metapkgs+=(./arch-packages/illogical-impulse-bibata-modern-classic-bin)
@@ -128,31 +113,9 @@ for i in "${metapkgs[@]}"; do
 	v install-local-pkgbuild "$i" "$metainstallflags"
 done
 
-# https://github.com/end-4/dots-hyprland/issues/428#issuecomment-2081690658
-# https://github.com/end-4/dots-hyprland/issues/428#issuecomment-2081701482
-# https://github.com/end-4/dots-hyprland/issues/428#issuecomment-2081707099
-case $SKIP_PYMYC_AUR in
-  true) sleep 0;;
-  *)
-	  pymycinstallflags=""
-	  $ask && showfun install-local-pkgbuild || pymycinstallflags="$pymycinstallflags --noconfirm"
-	  v install-local-pkgbuild "./arch-packages/illogical-impulse-pymyc-aur" "$pymycinstallflags"
-    ;;
-esac
-
-
-# Why need cleanbuild? see https://github.com/end-4/dots-hyprland/issues/389#issuecomment-2040671585
-# Why install deps by running a seperate command? see pinned comment of https://aur.archlinux.org/packages/hyprland-git
-case $SKIP_HYPR_AUR in
-  true) sleep 0;;
-  *)
-	  hyprland_installflags="-S"
-	  $ask || hyprland_installflags="$hyprland_installflags --noconfirm"
-    v yay $hyprland_installflags --asdeps hyprutils-git hyprlang-git hyprcursor-git hyprwayland-scanner-git
-    v yay $hyprland_installflags --answerclean=a hyprland-git
-    ;;
-esac
-
+# These python packages are installed using uv, not pacman.
+showfun install-python-packages
+v install-python-packages
 
 ## Optional dependencies
 if pacman -Qs ^plasma-browser-integration$ ;then SKIP_PLASMAINTG=true;fi
@@ -177,11 +140,9 @@ esac
 v sudo usermod -aG video,i2c,input "$(whoami)"
 v bash -c "echo i2c-dev | sudo tee /etc/modules-load.d/i2c-dev.conf"
 v systemctl --user enable ydotool --now
+v sudo systemctl enable bluetooth --now
 v gsettings set org.gnome.desktop.interface font-name 'Rubik 11'
-
-#####################################################################################
-printf "\e[36m[$0]: 2. Installing parts from source repo\e[0m\n"
-sleep 1
+v gsettings set org.gnome.desktop.interface color-scheme 'prefer-dark'
 
 if command -v LaTeX >/dev/null 2>&1;then
   echo -e "\e[33m[$0]: Program \"MicroTeX\" already exists, no need to install.\e[0m"
@@ -192,7 +153,7 @@ fi
 if $ask_MicroTeX;then showfun install-MicroTeX;v install-MicroTeX;fi
 
 #####################################################################################
-printf "\e[36m[$0]: 3. Copying + Configuring\e[0m\n"
+printf "\e[36m[$0]: 2. Copying + Configuring\e[0m\n"
 
 # In case some folders does not exists
 v mkdir -p $XDG_BIN_HOME $XDG_CACHE_HOME $XDG_CONFIG_HOME $XDG_DATA_HOME
@@ -226,15 +187,15 @@ esac
 case $SKIP_AGS in
   true) sleep 0;;
   *)
-    v rsync -av --delete --exclude '/user_options.js' .config/ags/ "$XDG_CONFIG_HOME"/ags/
-    t="$XDG_CONFIG_HOME/ags/user_options.js"
+    v rsync -av --delete --exclude '/user_options.jsonc' .config/ags/ "$XDG_CONFIG_HOME"/ags/
+    t="$XDG_CONFIG_HOME/ags/user_options.jsonc"
     if [ -f $t ];then
       echo -e "\e[34m[$0]: \"$t\" already exists.\e[0m"
-      # v cp -f .config/ags/user_options.js $t.new
+      # v cp -f .config/ags/user_options.jsonc $t.new
       existed_ags_opt=y
     else
       echo -e "\e[33m[$0]: \"$t\" does not exist yet.\e[0m"
-      v cp .config/ags/user_options.js $t
+      v cp .config/ags/user_options.jsonc $t
       existed_ags_opt=n
     fi
     ;;
@@ -244,16 +205,43 @@ esac
 case $SKIP_HYPRLAND in
   true) sleep 0;;
   *)
-    v rsync -av --delete --exclude '/custom' --exclude '/hyprland.conf' .config/hypr/ "$XDG_CONFIG_HOME"/hypr/
+    v rsync -av --delete --exclude '/custom' --exclude '/hyprlock.conf' --exclude '/hypridle.conf' --exclude '/hyprland.conf' .config/hypr/ "$XDG_CONFIG_HOME"/hypr/
     t="$XDG_CONFIG_HOME/hypr/hyprland.conf"
     if [ -f $t ];then
       echo -e "\e[34m[$0]: \"$t\" already exists.\e[0m"
-      v cp -f .config/hypr/hyprland.conf $t.new
-      existed_hypr_conf=y
+      if [ -f "$XDG_STATE_HOME/ags/user/firstrun.txt" ]
+      then
+        v cp -f .config/hypr/hyprland.conf $t.new
+        existed_hypr_conf=y
+      else
+        v mv $t $t.old
+        v cp -f .config/hypr/hyprland.conf $t
+        existed_hypr_conf_firstrun=y
+      fi
     else
       echo -e "\e[33m[$0]: \"$t\" does not exist yet.\e[0m"
       v cp .config/hypr/hyprland.conf $t
       existed_hypr_conf=n
+    fi
+    t="$XDG_CONFIG_HOME/hypr/hypridle.conf"
+    if [ -f $t ];then
+      echo -e "\e[34m[$0]: \"$t\" already exists.\e[0m"
+      v cp -f .config/hypr/hypridle.conf $t.new
+      existed_hypridle_conf=y
+    else
+      echo -e "\e[33m[$0]: \"$t\" does not exist yet.\e[0m"
+      v cp .config/hypr/hypridle.conf $t
+      existed_hypridle_conf=n
+    fi
+    t="$XDG_CONFIG_HOME/hypr/hyprlock.conf"
+    if [ -f $t ];then
+      echo -e "\e[34m[$0]: \"$t\" already exists.\e[0m"
+      v cp -f .config/hypr/hyprlock.conf $t.new
+      existed_hyprlock_conf=y
+    else
+      echo -e "\e[33m[$0]: \"$t\" does not exist yet.\e[0m"
+      v cp .config/hypr/hyprlock.conf $t
+      existed_hyprlock_conf=n
     fi
     t="$XDG_CONFIG_HOME/hypr/custom"
     if [ -d $t ];then
@@ -269,9 +257,6 @@ esac
 # some foldes (eg. .local/bin) should be processed separately to avoid `--delete' for rsync,
 # since the files here come from different places, not only about one program.
 v rsync -av ".local/bin/" "$XDG_BIN_HOME"
-
-# Dark mode by default
-v gsettings set org.gnome.desktop.interface color-scheme 'prefer-dark'
 
 # Prevent hyprland from not fully loaded
 sleep 1
@@ -296,8 +281,8 @@ done
 #####################################################################################
 printf "\e[36m[$0]: Finished. See the \"Import Manually\" folder and grab anything you need.\e[0m\n"
 printf "\n"
-printf "\e[36mIf you are new to Hyprland, please read\n"
-printf "https://end-4.github.io/dots-hyprland-wiki/en/i-i/01setup/#post-installation\n"
+printf "\e[36mIt is recommended to check out\n"
+printf "https://end-4.github.io/dots-hyprland-wiki/en/i-i/01setup/#post-installation \n"
 printf "for hints on launching Hyprland.\e[0m\n"
 printf "\n"
 printf "\e[36mIf you are already running Hyprland,\e[0m\n"
@@ -306,14 +291,25 @@ printf "\e[36mPress \e[30m\e[46m Super+/ \e[0m\e[36m for a list of keybinds\e[0m
 printf "\n"
 
 case $existed_ags_opt in
-  y) printf "\n\e[33m[$0]: Warning: \"$XDG_CONFIG_HOME/ags/user_options.js\" already existed before and we didn't overwrite it. \e[0m\n"
-#    printf "\e[33mPlease use \"$XDG_CONFIG_HOME/ags/user_options.js.new\" as a reference for a proper format.\e[0m\n"
+  y) printf "\n\e[33m[$0]: Warning: \"$XDG_CONFIG_HOME/ags/user_options.jsonc\" already existed before and we didn't overwrite it. \e[0m\n"
+#    printf "\e[33mPlease use \"$XDG_CONFIG_HOME/ags/user_options.jsonc.new\" as a reference for a proper format.\e[0m\n"
 ;;esac
-case $existed_hypr_conf in
-  y) printf "\n\e[33m[$0]: Warning: \"$XDG_CONFIG_HOME/hypr/hyprland.conf\" already existed before and we didn't overwrite it. \e[0m\n"
-     printf "\e[33mPlease use \"$XDG_CONFIG_HOME/hypr/hyprland.conf.new\" as a reference for a proper format.\e[0m\n"
-     printf "\e[33mIf this is your first time installation, you must overwrite \"$XDG_CONFIG_HOME/hypr/hyprland.conf\" with \"$XDG_CONFIG_HOME/hypr/hyprland.conf.new\".\e[0m\n"
+case $existed_hypr_conf_firstrun in
+  y) printf "\n\e[33m[$0]: Warning: \"$XDG_CONFIG_HOME/hypr/hyprland.conf\" already existed before. As it seems it is your firstrun, we replaced it with a new one. \e[0m\n"
+     printf "\e[33mAs it seems it is your first run, we replaced it with a new one. The old one has been renamed to \"$XDG_CONFIG_HOME/hypr/hyprland.conf.old\".\e[0m\n"
 ;;esac
+case $existed_hypridle_conf in
+  y) printf "\n\e[33m[$0]: Warning: \"$XDG_CONFIG_HOME/hypr/hypridle.conf\" already existed before and we didn't overwrite it. \e[0m\n"
+     printf "\e[33mPlease use \"$XDG_CONFIG_HOME/hypr/hypridle.conf.new\" as a reference for a proper format.\e[0m\n"
+;;esac
+case $existed_hyprlock_conf in
+  y) printf "\n\e[33m[$0]: Warning: \"$XDG_CONFIG_HOME/hypr/hyprlock.conf\" already existed before and we didn't overwrite it. \e[0m\n"
+     printf "\e[33mPlease use \"$XDG_CONFIG_HOME/hypr/hyprlock.conf.new\" as a reference for a proper format.\e[0m\n"
+;;esac
+
+if [[ -z "${ILLOGICAL_IMPULSE_VIRTUAL_ENV}" ]]; then
+  printf "\n\e[31m[$0]: \!! Important \!! : Please ensure environment variable \e[0m \$ILLOGICAL_IMPULSE_VIRTUAL_ENV \e[31m is set to proper value (by default \"~/.local/state/ags/.venv\"), or AGS config will not work. We have already provided this configuration in ~/.config/hypr/hyprland/env.conf, but you need to ensure it is included in hyprland.conf, and also a restart is needed for applying it.\e[0m\n"
+fi
 
 if [[ ! -z "$warn_files" ]]; then
   printf "\n\e[31m[$0]: \!! Important \!! : Please delete \e[0m ${warn_files[*]} \e[31m manually as soon as possible, since we\'re now using AUR package or local PKGBUILD to install them for Arch(based) Linux distros, and they'll take precedence over our installation.\e[0m\n"
